Fix ending condition.

+++ dqn_atari.py
@@ -488,6 +488,8 @@ def main():  # noqa: D103
             replay_memory.append_all(worker_memories)
             append_time += time.time() - start_time
             total_appended += len(worker_memories)
+            if total_appended > 100000:
+              break
             if args.psc:
                 os.remove(memory_path)
